+++ packages/orq-rc/src/models/operations/listbudgets.ts
@@ -474,7 +474,7 @@ export const ListBudgetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-04T13:27:51.444Z",
+    "2025-11-05T05:24:55.569Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -509,7 +509,7 @@ export const ListBudgetsData$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => ListBudgetsConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-04T13:27:51.444Z"))
+  updated: z.date().default(() => new Date("2025-11-05T05:24:55.569Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {